server/requirements.txt
algoliasearch==1.18.1
atomicwrites==1.1.5
attrs==18.1.0
bcrypt==3.1.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
Flask==1.0
Flask-Bcrypt==0.7.1
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
get==1.0.3
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
more-itertools==4.3.0
pluggy==0.7.1
post==1.0.2
psycopg2==2.7.3
public==1.0.3
py==1.5.4
pycparser==2.18
PyJWT==1.4.2
pytest==3.7.1
pytest-flask==0.10.0
python-http-client==3.1.0
query-string==1.0.2
request==1.0.2
requests==2.19.1
sendgrid==5.4.1
six==1.11.0
SQLAlchemy==1.1.13
urllib3==1.23
Werkzeug==0.14.1
